# Enhancement and Fixes in GPU Tsit5