--- a/.gitignore
+++ b/.gitignore
@@ -1,6 +1,7 @@
 *.jl.*.cov
 *.jl.cov
 *.jl.mem
+*.jl.*.mem
 .DS_Store
 /Manifest.toml
 /dev/