--- a/src/DiffEqGPU.jl
+++ b/src/DiffEqGPU.jl
@@ -172,7 +172,7 @@ function EnsembleGPUArray()
 end
 
 function EnsembleGPUKernel()
-    EnsembleGPUKernel(0.0)
+    EnsembleGPUKernel(0.2)
 end
 
 function ChainRulesCore.rrule(::Type{<:EnsembleGPUArray})
@@ -327,16 +327,17 @@ function batch_solve(ensembleprob, alg,
                 ts, us = vectorized_solve(cu(probs), ensembleprob.prob, GPUSimpleTsit5();
                                           kwargs...)
             end
-            solus = Array(us)
-            solts = Array(ts)
-            [@views ensembleprob.output_func(SciMLBase.build_solution(probs[i], alg,
-                                                                      solts[:, i],
-                                                                      solus[:, i],
-                                                                      k = nothing,
-                                                                      destats = nothing,
-                                                                      calculate_error = false),
-                                             i)[1]
-             for i in eachindex(probs)]
+            []
+            # solus = Array(us)
+            # solts = Array(ts)
+            # [@views ensembleprob.output_func(SciMLBase.build_solution(probs[i], alg,
+            #                                                           solts[:, i],
+            #                                                           solus[:, i],
+            #                                                           k = nothing,
+            #                                                           destats = nothing,
+            #                                                           calculate_error = false),
+            #                                  i)[1]
+            #  for i in eachindex(probs)]
 
         else
             error("We don't have solvers implemented for this algorithm yet")