--- a/src/gpu_tsit5.jl
+++ b/src/gpu_tsit5.jl
@@ -107,7 +107,7 @@ end
 end
 ## GPU solver
 
-function vectorized_solve(prob::ODEProblem, ps::CuVector, alg::GPUSimpleTsit5;
+function vectorized_solve(probs, prob::ODEProblem, alg::GPUSimpleTsit5;
                           dt, saveat = nothing,
                           save_everystep = true,
                           debug = false, kwargs...)
@@ -119,15 +119,15 @@ function vectorized_solve(prob::ODEProblem, ps::CuVector, alg::GPUSimpleTsit5;
         else
             len = 2
         end
-        ts = CuMatrix{typeof(dt)}(undef, (len, length(ps)))
-        us = CuMatrix{typeof(prob.u0)}(undef, (len, length(ps)))
+        ts = CuMatrix{typeof(dt)}(undef, (len, length(probs)))
+        us = CuMatrix{typeof(prob.u0)}(undef, (len, length(probs)))
     else
         error("Not fully implemented yet")  # see the TODO in the kernel
         ts = saveat
         us = CuMatrix{typeof(prob.u0)}(undef, (length(ts), length(ps)))
     end
 
-    kernel = @cuda launch=false tsit5_kernel(prob, ps, us, ts, dt,
+    kernel = @cuda launch=false tsit5_kernel(probs, us, ts, dt,
                                              Val(saveat !== nothing), Val(save_everystep))
     if debug
         @show CUDA.registers(kernel)
@@ -135,11 +135,11 @@ function vectorized_solve(prob::ODEProblem, ps::CuVector, alg::GPUSimpleTsit5;
     end
 
     config = launch_configuration(kernel.fun)
-    threads = min(length(ps), config.threads)
+    threads = min(length(probs), config.threads)
     # XXX: this kernel performs much better with all blocks active
-    blocks = max(cld(length(ps), threads), config.blocks)
-    threads = cld(length(ps), blocks)
-    kernel(prob, ps, us, ts, dt; threads, blocks)
+    blocks = max(cld(length(probs), threads), config.blocks)
+    threads = cld(length(probs), blocks)
+    kernel(probs, us, ts, dt; threads, blocks)
 
     # we build the actual solution object on the CPU because the GPU would create one
     # containig CuDeviceArrays, which we cannot use on the host (not GC tracked,
@@ -203,14 +203,15 @@ end
 # saveat is just a bool here:
 #  true: ts is a vector of timestamps to read from
 #  false: each ODE has its own timestamps, so ts is a vector to write to
-function tsit5_kernel(_prob, ps, _us, _ts, dt,
+function tsit5_kernel(probs, _us, _ts, dt,
                       ::Val{saveat}, ::Val{save_everystep}) where {saveat, save_everystep}
     i = (blockIdx().x - 1) * blockDim().x + threadIdx().x
-    i >= length(ps) && return
+    i >= length(probs) && return
 
     # get the actual problem for this thread
-    p = @inbounds ps[i]
-    prob = remake(_prob; p)
+    # p = @inbounds ps[i]
+    prob = @inbounds probs[i]
+    #prob = remake(_prob; p)
 
     # get the input/output arrays for this thread
     ts = if saveat