--- a/test/gpu_tsit5_tests.jl
+++ b/test/gpu_tsit5_tests.jl
@@ -16,9 +16,9 @@ prob = ODEProblem{false}(lorenz, u0, tspan, p)
 prob_func = (prob, i, repeat) -> remake(prob, p = p)
 monteprob = EnsembleProblem(prob, prob_func = prob_func, safetycopy = false)
 
-sol = solve(monteprob, GPUTsit5(), EnsembleGPUAutonomous(), trajectories = 10,
+sol = solve(monteprob, GPUTsit5(), EnsembleGPUKernel(), trajectories = 10,
             adaptive = false, dt = 0.1f0)
-asol = solve(monteprob, GPUTsit5(), EnsembleGPUAutonomous(), trajectories = 10,
+asol = solve(monteprob, GPUTsit5(), EnsembleGPUKernel(), trajectories = 10,
              adaptive = true, dt = 0.1f-1, abstol = 1.0f-8, reltol = 1.0f-5)
 
 @test sol.converged == true