Hi,

Wrt the previous discussion on speed-up issues with DiffEqGPU API with the new GPU solvers, I set to bridge the performance gap (46 µs to 10.5 ms 😱) between raw GPU solutions and using the Ensemble DiffEqGPU API. A small commit fix improved the speed by ~ 2.4 ms. (The u0 and p `hcat` was only needed by previous solving techniques).
Pros: It is now faster than `EnsembleGPUArray` and `EnsembleCPUArray`.

Before:

```
julia> @benchmark sol = solve(monteprob, GPUTsit5(), EnsembleGPUAutonomous(), trajectories = 1000,
                              adaptive = false, dt = dt)

BenchmarkTools.Trial: 386 samples with 1 evaluation.
 Range (min … max):   8.805 ms … 40.638 ms  ┊ GC (min … max): 0.00% … 23.29%
 Time  (median):     10.538 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.962 ms ±  5.690 ms  ┊ GC (mean ± σ):  8.49% ± 10.88%

  ▄▂▁▄█▄▂ ▁                                    ▁               
  █████████▇▇▇▇▄▇▇▇▄▄▄▆▁▅▁▁▁▁▅▄▁▅▄▁▁▁▁▁▄▆▅▁▄▅▆▇██▅▅▆▁▁▄▁▁▄▅▁▅ ▆
  8.81 ms      Histogram: log(frequency) by time      30.3 ms <

 Memory estimate: 13.86 MiB, allocs estimate: 88075.
```

After:

```
julia> @benchmark sol = solve(monteprob, GPUTsit5(), EnsembleGPUAutonomous(), trajectories = 1000,
                              adaptive = false, dt = dt)
BenchmarkTools.Trial: 497 samples with 1 evaluation.
 Range (min … max):   7.095 ms … 74.585 ms  ┊ GC (min … max): 0.00% … 24.81%
 Time  (median):      8.108 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   10.051 ms ±  7.866 ms  ┊ GC (mean ± σ):  8.66% ±  9.01%

  ▂█▅ ▁                                                        
  ███▄█▅▁▁▁▄▁▁▁▁▁▁▁▄▁▄▁▁▁▁▁▁▄▁▁▄▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▄▁▄▅▁▇▇ ▆
  7.09 ms      Histogram: log(frequency) by time      43.3 ms <

 Memory estimate: 5.88 MiB, allocs estimate: 76811.
```

With some profiling and tracking allocations (attached in the next comment), I was able to lower these issues:

- [x] remove `u0` and `p` build from `EnsembleGPUAutonomous`
- [x] Search for better alternatives for building `u0` and `p` in `EnsembleGPUArray`
- [x] Check for `@assert` necessity
- [ ] Transforming GPU `us` arrays to CPU arrays
- [x] Possibility of building Vector{ODESolution} with better speed and lower allocations

@ChrisRackauckas, have a look.