@ChrisRackauckas 2022-07-27
Why does that allocate? Is there a way using `Fix2` to make it not allocate?