@utkarsh530 2022-07-31
Made a fix with `Fix2`. It is now reduced to 2 allocations.

Some speed-ups as well:


Before:

```
julia> @benchmark sol = solve(monteprob, GPUTsit5(), EnsembleGPUAutonomous(), trajectories = 1000,
                              adaptive = false, dt = dt)
BenchmarkTools.Trial: 497 samples with 1 evaluation.
 Range (min … max):   7.095 ms … 74.585 ms  ┊ GC (min … max): 0.00% … 24.81%
 Time  (median):      8.108 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   10.051 ms ±  7.866 ms  ┊ GC (mean ± σ):  8.66% ±  9.01%

  ▂█▅ ▁                                                        
  ███▄█▅▁▁▁▄▁▁▁▁▁▁▁▄▁▄▁▁▁▁▁▁▄▁▁▄▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▄▁▄▅▁▇▇ ▆
  7.09 ms      Histogram: log(frequency) by time      43.3 ms <

 Memory estimate: 5.88 MiB, allocs estimate: 76811.

```

After:

```
julia> @benchmark solve(monteprob, GPUTsit5(), EnsembleGPUAutonomous(), trajectories = 1000,
                   adaptive = false, dt = 0.1f0)
BenchmarkTools.Trial: 3789 samples with 1 evaluation.
 Range (min … max):  802.894 μs … 44.251 ms  ┊ GC (min … max):  0.00% … 61.15%
 Time  (median):     875.493 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):     1.311 ms ±  2.630 ms  ┊ GC (mean ± σ):  26.85% ± 14.05%

  █▄▂▁                                                          
  ████▅▄▃▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▃▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄█▇▆▄▅▁▄ █
  803 μs        Histogram: log(frequency) by time      12.7 ms <

 Memory estimate: 3.86 MiB, allocs estimate: 34400.
```

We are now 43x faster than `EnsembleGPUArray` and 14x faster than `EnsembleCPUArray` .
Moreover, now the current implementation supports parameter parallelism across `u0` as well instead of `p`.