@ChrisRackauckas 2022-08-01
And what's the overhead vs the most direct form, and the current reason for the overhead?