@utkarsh530 2022-08-01
> And what's the overhead vs. the most direct form, and the current reason for the overhead?

The major overhead is converting back to CPU Arrays, `probs` creation and finally building the solution.
<img width="1033" alt="Screenshot 2022-08-02 at 12 09 10 AM" src="https://user-images.githubusercontent.com/37050056/182220489-9e82dcf8-2b3c-41d0-a876-e413d96468e5.png">