@utkarsh530 2022-08-02
Cross-posted from slack:

> And what’s the overhead vs. the most direct form, and the current reason for the overhead?

The reason for this overhead  (converting to CPU Arrays) is to provide users to access something like sol[i].u[j] where i,j are some indexes. It would cause scalar indexing on ts,us  which are CUArrays.
I wanted to ask whether should we leave it user to convert to CPU Arrays? I missed that thing to discuss.
The probs  creation seems to be necessary, but maybe could be pulled out of DiffEqGPU? Currently, it was done to adhere to DiffEqGPU way of handling it.  This was not coming in the previous benchmarks because ps was being built separately and passed to the vectorized_solve .
And sorry for some sloppiness which may had caused any discomfort.