@ChrisRackauckas 2022-08-02
no

> what does cu(probs) actually do? What dispatch is that hitting?